tvm_ffi::dtype

Trait AsDLDataType

pub trait AsDLDataType: Copy {
    const DL_DATA_TYPE: DLDataType;
}

Trait to convert standard data types to DLDataType

This trait provides a way to get the corresponding DLDataType for standard Rust types. It’s implemented for common integer, unsigned integer, and floating-point types.

Required Associated Constants

const DL_DATA_TYPE: DLDataType

The corresponding DLDataType for this type

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.