tvm_ffi

Module error

Structs

Error

Error reference class

ErrorKind

Error kind, wraps in a struct to be explicit

ErrorObj

error object

Constants

ATTRIBUTE_ERROR

INDEX_ERROR

KEY_ERROR

RUNTIME_ERROR

TYPE_ERROR

VALUE_ERROR

Type Aliases

Result

Default result that uses Error as the error type