tvm_ffi::error

Struct Error

pub struct Error { /* private fields */ }

Error reference class

Implementations

impl Error

pub fn new(kind: ErrorKind<'_>, message: &str, traceback: &str) -> Self

pub fn from_raised() -> Self

Create a new error by moving from raised error

Returns

The error from the raised error

pub fn set_raised(error: &Self)

Set the error as raised

Arguments

error - The error to set as raised

pub fn kind(&self) -> ErrorKind<'_>

Get the kind of the error

Returns

The kind of the error

pub fn message(&self) -> &str

Get the message of the error

Returns

The message of the error

pub fn backtrace(&self) -> &str

Get the backtrace of the error

Returns

The backtrace of the error

pub fn traceback_most_recent_call_last(&self) -> String

Get the traceback of the error in the order of most recent call last

Returns

The traceback of the error

pub fn with_appended_backtrace(this: Self, backtrace: &str) -> Self

Append the backtrace to the error

Arguments

this - The error to append the backtrace to

backtrace - The backtrace to append

Returns

The error with the appended backtrace