tvm_ffi::function

Struct Function

pub struct Function { /* private fields */ }

Error reference class

Implementations

impl Function

pub fn call_packed(&self, packed_args: &[AnyView<'_>]) -> Result<Any>

Call the function in packed format.

pub fn call_tuple<TupleType>(&self, tuple_args: TupleType) -> Result<Any>
where TupleType: TupleAsPackedArgs,

pub fn call_tuple_with_len<const LEN: usize, TupleType>( &self, tuple_args: TupleType, ) -> Result<Any>
where TupleType: TupleAsPackedArgs,

Call function with compile-time known argument count This is an optimized version of call_tuple for when the argument count is known at compile time, avoiding the small vector optimization overhead.

Arguments

tuple_args - The tuple arguments

Returns

Any - The result

pub fn get_global(name: &str) -> Result<Function>

Get global function by name This function will throw an error if the function is not found.

Arguments

name - The name of the function

Returns

Function - The global function

pub fn register_global(name: &str, func: Function) -> Result<()>

Register a function as a global function

Arguments

name - The name of the function

func - The function to register

Returns

Result<()> - The result of the registration

pub fn from_packed<F>(func: F) -> Self
where F: Fn(&[AnyView<'_>]) -> Result<Any> + 'static,

Construct a function from a packed function

Arguments

func - The packed function in signature of Fn(&[AnyView]) -> Result<Any>

Returns

Function - The function

pub fn from_typed<F, I, O>(func: F) -> Self
where F: AsPackedCallable<I, O> + 'static,

Construct a function from a typed function

Arguments

func - The typed function with function signature of F(T0, T1, ...) -> Result<O>

Returns

Function - The function

pub fn from_extern_c( handle: *mut c_void, safe_call: TVMFFISafeCallType, deleter: Option<unsafe extern "C" fn(*mut c_void)>, ) -> Self