tvm_ffi

Macro into_typed_fn

macro_rules! into_typed_fn {
    ($f:expr, $trait:ident() -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty) -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty) -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty, $t2:ty) -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty, $t2:ty, $t3:ty) -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty, $t2:ty, $t3:ty, $t4:ty) -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty, $t2:ty, $t3:ty, $t4:ty, $t5:ty) -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty, $t2:ty, $t3:ty, $t4:ty, $t5:ty, $t6:ty)
        -> $ret_ty:ty) => { ... };
    ($f:expr, $trait:ident($t0:ty, $t1:ty, $t2:ty, $t3:ty, $t4:ty, $t5:ty, $t6:ty, $t7:ty)
        -> $ret_ty:ty) => { ... };
}

into_typed_fn

Converts a generic Function into a typed function with compile-time argument count and type checking. This macro provides a convenient way to create type-safe wrappers around TVM functions.

Arguments

$f - The function identifier to convert

$trait - The trait type (typically Fn)

($t0, $t1, ...) - The argument types

$ret_ty - The return type

Example

use tvm_ffi::*;

let func = Function::from_typed(|x: i32, y: i32| -> Result<i32> { Ok(x + y) });
let typed_func = into_typed_fn!(func, Fn(i32, &i32) -> Result<i32>);
let result = typed_func(10, &20).unwrap(); // Returns 30
assert_eq!(result, 30);

Note that the into_typed_fn! macro can specify arguments to be passed either by reference or by value in the argument list. We recommend passing by reference for ObjectRef types such as Tensor. Since the ffi mechanism requires us to pass arguments by reference.

Supported Argument Counts

This macro supports functions with 0 to 8 arguments.