tvm_ffi::object

Trait ObjectCore

pub unsafe trait ObjectCore: Sized + 'static {
    const TYPE_KEY: &'static str;

    // Required methods
    fn type_index() -> i32;
    unsafe fn object_header_mut(this: &mut Self) -> &mut TVMFFIObject;
}

Traits that can be used to check if a type is an object

This trait is unsafe because it is used to access the object header and the object header is unsafe to access

Required Associated Constants

const TYPE_KEY: &'static str

the type key of the object

Required Methods

fn type_index() -> i32

unsafe fn object_header_mut(this: &mut Self) -> &mut TVMFFIObject

Return the object header This function is implemented as a static function so

Arguments

this - The object to get the header

Returns

&mut TVMFFIObject - The object header \return The object header

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.