tvm_ffi::type_traits

Trait AnyCompatible

pub unsafe trait AnyCompatible: Sized {
    // Required methods
    unsafe fn copy_to_any_view(src: &Self, data: &mut TVMFFIAny);
    unsafe fn move_to_any(src: Self, data: &mut TVMFFIAny);
    unsafe fn check_any_strict(data: &TVMFFIAny) -> bool;
    unsafe fn copy_from_any_view_after_check(data: &TVMFFIAny) -> Self;
    unsafe fn move_from_any_after_check(data: &mut TVMFFIAny) -> Self;
    unsafe fn try_cast_from_any_view(data: &TVMFFIAny) -> Result<Self, ()>;
    fn type_str() -> String;

    // Provided method
    fn get_mismatch_type_info(data: &TVMFFIAny) -> String { ... }
}

Trait to enable a value to be compatible with Any Enables TryFrom/Into AnyView/Any

Required Methods

unsafe fn copy_to_any_view(src: &Self, data: &mut TVMFFIAny)

the value to copy to TVMFFIAny

unsafe fn move_to_any(src: Self, data: &mut TVMFFIAny)

consume the value to move to Any

unsafe fn check_any_strict(data: &TVMFFIAny) -> bool

unsafe fn copy_from_any_view_after_check(data: &TVMFFIAny) -> Self

Copy value from TVMFFIAny after checking caller must ensure that the value is compatible with the type

unsafe fn move_from_any_after_check(data: &mut TVMFFIAny) -> Self

the value to move from TVMFFIAny NOTE: pay very careful attention to avoid memory leak!

When calling from managed Any, remember to use std::mem::ManuallyDrop

unsafe fn try_cast_from_any_view(data: &TVMFFIAny) -> Result<Self, ()>

try to cast the value from AnyView

fn type_str() -> String

the type string of the type

Provided Methods

fn get_mismatch_type_info(data: &TVMFFIAny) -> String

Get the type key of a type when TryCastFromAnyView fails.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.